+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - avoiding check
 - figure color to short name
 - readme
+- castling when is check for king is forbidden
 
 ## [0.3.3] - 2019-01-17
 ### Added